Extract Transliteration to a new file
